Fix crash when `bundle exec`'ing to bundler
I can't think of any reason to `bundle exec`'ing to `bundler`, but it
shouldn't be a way to make bundler crash.

In this case, the crash was due to `bundle exec`'ing on a definition
missing the current specific platform (due to lockfile generated with
older bundler). The definition would resolve gems in the main process,
and then would tried to be re-resolved again inside the "bundle exec"
context because of the missing platform. This double resolution would
cause bundler to crash.

I don't see why we should raise an exception in this situation,
specially such a cryptic one, so don't raise it.